Moved lbBindHtml in a separate directive
